#1134 `/execution-plans`-based api parts moved to ExecutionPlansController (changed from @Controller to @RestController). Changes successfully tested in postman.
